More code to the qspnclient.
The program maintains some data related to an identity, that is to a
NodeID. These are 'ready', 'my_naddr', 'my_fp' and now also
'addr_man'. Move that data to the class ''IdentityData'' internal to
the program.
This removes the need to put them in the old class ''QspnInitData''
which was saved inside the IdentityManager.
Implement 'get_identity_skeleton' and 'get_identity_skeleton_set'.