Improvement to qspnclient.
As per documentation improvement, for any destination we have to add
routes towards every internal IP address that is valid.
This commit updates the functions which are used to compute the IP
addresses (particularly the one to compute iternal IP that now
receives one more parameter). Further, all the cases are more
accurately invetigated when a new destination (not a path) is
notified.
The changes during the choice of the best path (method
`update_best_path`) is yet to be completed.