Send message "destroy" to neighbors.
When a system exits the program, every identity of the system
sends the message "destroy" to all its neighbors.

When a g-node *w* enters another network, it is immediately
removed from the old network. Every old identity sends the message
"destroy" to the neighbors which were not in *w*.

A particular case is when a g-node *w* migrates from a g-node *g*
into a g-node *h* in the same network.

In this case the old g-node can be removed only after verification
that *g* is internally connected. This verification is carried on
by a single node; thus, when the verification succeeds the
information has to be propagated to all the nodes of *w*.

Only after that (signal `remove_identity`), every old identity sends
the message "destroy" to the neighbors which were not in *w*.